Add Runner class
Runner is the main interface to parallel.  It takes a number of workers
to be used and will initialize as many workers as wanted.  Workers will
pick up tasks as soon as they arrive and the results will be available
in the output queue, also accessible through the results() method.